Crate anvil_zksync_traces

Modules

abi_utils

ABI related helper functions.

decode

format

identifier

writer

Helper methods to display transaction data in more human readable way.

Functions

build_call_trace_arena

Builds a call trace arena from the calls and transaction result.

decode_trace_arena

Decode a collection of call traces.

filter_call_trace_arena

Filter a call trace arena based on verbosity level.

render_trace_arena_inner

Render a collection of call traces to a string