anvil_zksync_traces

Module decode

Modules

revert_decoder

Various utilities to decode test results.

Structs

CallTraceDecoder

The call trace decoder.

CallTraceDecoderBuilderBase

A base struct for builders. If you want to start building from a predefined CallTraceDecoder instance, create your own builder like that:

Constants

SELECTOR_LEN

The first four bytes of the call data for a function call specifies the function to be called.

Functions

decode_value

get_indexed_event_from_vm_event

Given an Event without indexed parameters and a VmEvent, it tries to return the Event with the proper indexed parameters. Otherwise, it returns the original Event.

label_value

vm_event_to_log_data

Converts a VmEvent to a LogData.